+++ CHANGES.rst
@@ -23,7 +23,11 @@ Unreleased
   that missing branches are reported near the other lines they affect. The
   values used to show all missing lines, and then all missing branches.
 
+- Line numbers in the HTML report now align properly with source lines, even
+  when Chrome's minimum font size is set, fixing `issue 748`_.  Thanks Wen Ye.
+
 .. _issue 746: https://github.com/nedbat/coveragepy/issues/746
+.. _issue 748: https://github.com/nedbat/coveragepy/issues/748
 
 
 .. _changes_50a4: